+++ changelog.md
@@ -1,8 +1,20 @@
 # Changelog
 
+**Test Status**
+* Alarm API test results match between both controller builds.
+
+---
+
+**./test-parts/part-h-api_requests/api-main.js**
+* Uncommented 'alarmTests' call.
+
+---
+
 **./test-parts/part-h-api_requests/requests/req-admin_health.js**
-* Moved functions:
-	* 'testDriveLetter' as 'testDriveLetterString'
-	* 'testNetworkInternal' as 'testNetworkInternalArray'
-* Destination:
-	* ../sub-requests/common-storage.js
+* Restored the 'testDriveLetter' function.
+
+---
+
+**./test-parts/part-h-api_requests/sub-requests/common-health.js**
+* Moved the 'testDriveLetterString' function back to:
+	* ../requests/req-admin_health.js